refactor: перенести вспомогательные функции в общий файл
