--- src/utils/getAge.js
@@ -1,9 +1,0 @@
-export const getAge = (date) => {
-  const today = new Date();
-  let age = today.getFullYear() - date.getFullYear();
-  const month = today.getMonth() - date.getMonth();
-  if (month < 0 || (month === 0 && today.getDate() < date.getDate())) {
-    age -= 1;
-  }
-  return age;
-};
